requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
cachetools==5.3.0
chardet==5.1.0
colorama==0.4.6
distlib==0.3.6
Django==4.2
django-filter==23.1
djangorestframework==3.14.0
exceptiongroup==1.1.1
filelock==3.11.0
flake8==6.0.0
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
packaging==23.0
platformdirs==3.2.0
pluggy==1.0.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
pyproject-api==1.5.1
pytest==7.3.0
pytest-django==4.5.2
pytz==2023.3
sqlparse==0.4.3
tomli==2.0.1
tox==4.4.11
virtualenv==20.21.0
